refactor: 直接dataモジュールに依存しないように
