+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/url/UrlPreviewStoreProvider.kt
@@ -6,26 +6,27 @@ import net.pantasystem.milktea.app_store.setting.SettingStore
 import net.pantasystem.milktea.data.infrastructure.url.db.UrlPreviewDAO
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.url.UrlPreviewStore
+import net.pantasystem.milktea.model.url.UrlPreviewStoreProvider
 import java.util.concurrent.ConcurrentHashMap
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class UrlPreviewStoreProvider @Inject constructor(
+class UrlPreviewStoreProviderImpl @Inject constructor(
     val settingStore: SettingStore,
     val urlPreviewDAO: UrlPreviewDAO,
     val accountStore: AccountStore,
     private val misskeyAPIServiceBuilder: MisskeyAPIServiceBuilder,
 
-    ) {
+    ) : UrlPreviewStoreProvider{
 
     private val mUrlPreviewStoreInstanceBaseUrlMap = ConcurrentHashMap<String, UrlPreviewStore>()
 
-    fun getUrlPreviewStore(account: Account): UrlPreviewStore {
+    override fun getUrlPreviewStore(account: Account): UrlPreviewStore {
         return getUrlPreviewStore(account, false)
     }
 
-    fun getUrlPreviewStore(
+    override fun getUrlPreviewStore(
         account: Account,
         isReplace: Boolean
     ): UrlPreviewStore {
