+++ modules/model/src/main/java/net/pantasystem/milktea/model/url/UrlPreviewStore.kt
@@ -1,6 +1,17 @@
 package net.pantasystem.milktea.model.url
 
+import net.pantasystem.milktea.model.account.Account
+
 interface UrlPreviewStore{
 
     fun get(url: String): UrlPreview?
+}
+
+interface UrlPreviewStoreProvider {
+    fun getUrlPreviewStore(account: Account): UrlPreviewStore
+
+    fun getUrlPreviewStore(
+        account: Account,
+        isReplace: Boolean
+    ): UrlPreviewStore
 }
